JOUR 2 - Correction CRUD produit complet

+++ SPRING-MVC/formation-spring-web/src/main/java/fr/formation/api/ProduitApiController.java
@@ -94,7 +94,14 @@ public class ProduitApiController {
 	@DeleteMapping("/{id}")
 	public boolean deleteById(@PathVariable int id) {
 		try {
+//			this.daoProduit.findById(id).orElseThrow(ProduitNotFoundException::new);
+			
+			if (!this.daoProduit.existsById(id)) {
+				throw new ProduitNotFoundException();
+			}
+			
 			this.daoProduit.deleteById(id);
+			
 			return true;
 		}
